requirements.txt
Brotli==1.1.0
pyahocorasick==2.1.0
pyparsing==3.2.3
tqdm==4.66.1
